Use async folder for universal components
