+++ client/async/HomePage.js
@@ -1,0 +1,1 @@
+export { default } from '../pages/Home/Home'
