+++ client/async/PostDetailsPage.js
@@ -1,0 +1,1 @@
+export { default } from '../pages/PostDetails/PostDetails'
