+++ client/async/PostsPage.js
@@ -1,0 +1,1 @@
+export { default } from '../pages/Posts/Posts'
