+++ client/async/UserCreatePostPage.js
@@ -1,0 +1,1 @@
+export { default } from '../pages/UserCreatePost/UserCreatePost'
