+++ client/async/UserPostsPage.js
@@ -1,0 +1,1 @@
+export { default } from '../pages/UserPosts/UserPosts'
