+++ client/async/UserProfilePage.js
@@ -1,0 +1,1 @@
+export { default } from '../pages/UserProfile/UserProfile'
