+++ client/async/UserRegisterPage.js
@@ -1,0 +1,1 @@
+export { default } from '../pages/UserRegister/UserRegister'
